Create Product component and shop page styling
The Product component is passed the info we get from the api call and
displays that information for each product. There are also buttons and
an input field to determine quantity for each product, which is handled
in state of the component. Also added stying to layout the products on
the shop page in a grid.